Align typeahead dropdown to the right of the content, and make sure it can flow across the editor
The typeahead is now added to the body tag, and not as a sibling to the tag where it is used. That way it can be placed on top of the rest of the page. We also let the dropdown take up as much width as it needs, to accommodate very long selections (death causes, I am looking at you!)