Base visibility of the 'Show in editor' link on a property from the backend, and not the current user
Before this change, we tested if the user was superuser, or if the user_id match the one on the post. Because it might be the case that a user can no longer edit their own posts (too much time has passed for instance). The check now lives in the backend. The frontend only listens for this bool and use that to show or hide the button.

https://trello.com/c/jrICqU7Y/91-fjernelse-af-superbruger-logik-pa-postvisning-vis-i-kildetaster