Make open typeahead/dropdown markup follow scroll
Because we are using a container with scroll, and not page scroll, the elements did not scroll when the content did. We now append them to the parent tag instead of the body. Because of this change, and because we work on a container with scroll, the content could be clipped. So instead of showing long items on a single line, we now break to allow content to be more lines.

Because of this, we adjust padding and line-height to make items that break easier to identify, and not confuse users to think they are two different items.

https://trello.com/c/XChGBuxj/142-felter-falder-ned-under-scrollbar