Output date of death in correct format
This commit includes a hack, because the data returned from the api is wrong (and resulted in the wrongfully formatted date).

We should use the data stored in the jsonObj property on each result, but for dateOfDeath we now copy that after having parsed the json. That way
we use the timestamp, and not the server formatted date (in the format of 04-06-1877)

The hack in searchresult.directive.js should be removed when the API returns the correct timestamp